Atoi

+++ pointers_arrays_strings/100-atoi.c
@@ -8,7 +8,7 @@ int _atoi(char *s)
 {
 	int i = 0;
 	int sign = 1;
-	int result = 0;
+	long int result = 0;
 
 	while (s[i] != '\0' && (s[i] < '0' || s[i] > '9'))
 	{
